% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_qc_to_prod}
\alias{compare_qc_to_prod}
\title{Compare production to QC versions of an SDTM table}
\usage{
compare_qc_to_prod(qc, prod, group_col)
}
\arguments{
\item{qc}{a data frame, the SDTM QC table}

\item{prod}{a data frame, the SDTM production table}

\item{group_col}{a character vector, the columns to use a unique keys for
comparing \code{qc} and \code{prod}. Ideally this is the key variables for the domain
as set out in the specification.}
}
\value{
a named list as returned by \code{compareDF::compare_df()}. A side effect
also prints a high level summary using \code{summary_diff()}.
}
\description{
Wraps \code{compareDF::compare_df()} function and prints a high level summary
using \code{\link{summary_diffs}}. The arguments of \code{compareDF::compare_df()}
are set to \code{stop_on_error = F} to avoid errors if the data frames are the
same, and \code{group_col} is set to \code{key_vars}.
}
