% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm}
\alias{dm}
\title{Example SDTM Domain 'DM'}
\format{
\subsection{'dm'}{

A data frame with 2 rows and 4 columns:
\describe{
\item{USUBJID}{Subject identifier}
\item{RFSTDTC}{Study start date}
\item{RFXSTDTC}{First exposure date}
\item{RFXENDTC}{Last exposure date}
}
}
}
\usage{
dm
}
\description{
This is an example data set to simulate a SDTM production domain 'DM' which
contains study start and end date information by subject. This can be used to
test \code{\link[=create_BLFL]{create_BLFL()}}, \code{\link[=create_EPOCH]{create_EPOCH()}}, and \code{\link[=calc_DY]{calc_DY()}}.
}
\keyword{datasets}
