% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edc_xx}
\alias{edc_xx}
\title{Example EDC data for form/table 'XX'}
\format{
\subsection{'edc_xx'}{

A data frame with 8 rows and 6 columns:
\describe{
\item{STUDYID}{Study identifier}
\item{USUBJID}{Subject identifier}
\item{VISIT}{Visit name}
\item{XXTESTCD}{Test name, coded}
\item{XXORRES}{Test result}
}
}
}
\usage{
edc_xx
}
\description{
This is an example data set to simulate raw EDC data from the fake form/table
'XX'.
}
\keyword{datasets}
