% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
\subsection{'example_data'}{

\describe{
\item{item 1}{Item 1 does this}
\item{item 2}{Item 2 does that}
}
}
}
\usage{
example_data
}
\description{
An example data set
}
\keyword{datasets}
