% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_info.R
\name{get_codelist}
\alias{get_codelist}
\title{Read in the code list from the specification for a specific domain}
\usage{
get_codelist(domain, dir_spec, filename_spec)
}
\arguments{
\item{domain}{string, SDTM domain or supplemental domain code}

\item{dir_spec}{string, specification directory}

\item{filename_spec}{string, file name of the specification}
}
\value{
a data frame with the code list
}
\description{
Reads-in the \code{"Codelists"} sheet from the study's specification MS Excel file
and then filters that code list by the variables in the domain
}
