% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_info.R
\name{get_data_spec}
\alias{get_data_spec}
\title{Read in the variable specification sheet for a SDTM data set}
\usage{
get_data_spec(domain, dir_spec, filename_spec)
}
\arguments{
\item{domain}{string, SDTM domain or supplemental domain code}

\item{dir_spec}{string, specification directory}

\item{filename_spec}{string, file name of the specification}
}
\value{
a data frame of the variable specification for \code{domain}
}
\description{
Reads the specified domain variable specification sheet from an MS Excel file.
}
\details{
The \code{read_excel()} function will causes an access denied warning when reading
in a read-only specification file. This does not affect the data import.
Variables will be arranged in descending order per the \code{"Order"} column in
the specification.
}
