% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_info.R
\name{get_key_vars}
\alias{get_key_vars}
\title{Read in the key variables for a SDTM domain}
\usage{
get_key_vars(domain, dir_spec, filename_spec)
}
\arguments{
\item{domain}{string, SDTM domain or supplemental domain code}

\item{dir_spec}{string, specification directory}

\item{filename_spec}{string, file name of the specification}
}
\value{
a character vector of key variables for the specified \code{domain}
}
\description{
Reads the \code{"Key Variables"} column from the SDTM specification MS Excel file's
\code{"Datasets"} sheet for the specified \code{domain}.
}
\details{
The \code{read_excel()} function will causes an access denied warning when reading
in a read-only specification file. This does not affect the data import.
}
