% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{get_valuelevel}
\alias{get_valuelevel}
\title{Read in value levels sheet from a study specification Excel file}
\usage{
get_valuelevel(
  domain,
  dir,
  filename,
  valuelevel_sheet = "ValueLevel",
  dataset_col = "Dataset"
)
}
\arguments{
\item{domain}{string, SDTM domain or supplemental domain code}

\item{dir}{string, specification directory}

\item{filename}{string, file name of the specification}

\item{valuelevel_sheet}{a string, the name of the value level sheet in
\code{filename}, default is \code{"ValueLevel"}}

\item{dataset_col}{a string, the name of the column in \code{valuelevelsheet}
for filtering my domain, default is \code{"Dataset"}}
}
\value{
a data frame with the value levels
}
\description{
Reads in the value levels sheet from a study specification Excel file and
filters the table for the entries relevant to \code{"domain"}
}
\examples{
work_dir <- system.file("extdata", package = "sdtmval")
get_valuelevel(domain = "SUPPXX",
               dir = work_dir,
               filename = "spec.xlsx")

}
\seealso{
\code{\link[=get_codelist]{get_codelist()}}, \code{\link[=get_data_spec]{get_data_spec()}}, \code{\link[=get_key_vars]{get_key_vars()}},
\code{\link[=assign_meta_data]{assign_meta_data()}}
}
