% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_codelists}
\alias{spec_codelists}
\title{Example 'Codelists' tab from a SDTM specification .xlsx file}
\format{
\subsection{'spec_codelists'}{

A data frame with 3 rows and 3 columns:
\describe{
\item{ID}{The variable identifier/name}
\item{Term}{The coded term}
\item{Decoded Value}{The corresponding decoded value for the coded term}
}
}
}
\usage{
spec_codelists
}
\description{
This table simulates an excerpt from a SDTM specification .xlsx file for the
'Codelists' tab which provides coded and decoded values from \code{XXTESTCD} and
\code{XXTEST} variables, respectively. This data set can be used to test the
\code{\link[=get_codelist]{get_codelist()}} function.
}
\keyword{datasets}
