% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_datasets}
\alias{spec_datasets}
\title{Example 'Datasets' tab from a SDTM specification .xlsx file}
\format{
\subsection{'spec_datasets'}{

A data frame with 1 row and 4 columns:
\describe{
\item{Dataset}{The domain}
\item{Description}{The domain description}
\item{Structure}{Defines what qualifies as a unique record}
\item{Key Variables}{The domain's key variables}
}
}
}
\usage{
spec_datasets
}
\description{
This table simulates an excerpt from a SDTM specification .xlsx file for the
'Datasets' tab which provides the key variables for the fake domain XX. This
data set can be used to test the \code{\link[=get_key_vars]{get_key_vars()}} function.
}
\keyword{datasets}
