% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_valuelevel}
\alias{spec_valuelevel}
\title{Example 'valuelevel' tab from a SDTM specification .xlsx file}
\format{
\subsection{'spec_valuelevel'}{

A data frame with 3 rows and 9 columns:
\describe{
\item{Dataset}{The dataset/domain}
\item{Variable}{The variable in dataset}
\item{Where Clause}{Applicable cases, in psuedocode}
\item{Description}{Label}
\item{Codelist}{Relevant codelist entries}
\item{Origin}{Where the data originated}
\item{Pages}{aCRF reference pages}
\item{Method}{Applicable methods}
\item{Programming Notes}{Programming notes, in psuedocode}
}
}
}
\usage{
spec_valuelevel
}
\description{
This table simulates an excerpt from a SDTM specification .xlsx file for the
'ValueLevel' tab. This data set can be used to test the \code{\link[=get_valuelevel]{get_valuelevel()}}
function.
}
\keyword{datasets}
