% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spec_XX}
\alias{spec_XX}
\title{Example domain specific tab from a SDTM specification .xlsx file}
\format{
\subsection{'spec_XX'}{

A data frame with 12 rows and 5 columns:
\describe{
\item{Order}{The order of the varibles in the data set}
\item{Dataset}{The domain abbreviation}
\item{Varible}{The domain's variables}
\item{Label}{Variable labels}
\item{Data Type}{Variable data types}
\item{Length}{The maximum allowed length of an entry}
}
}
}
\usage{
spec_XX
}
\description{
This table simulates an excerpt from a SDTM specification .xlsx file for the
fake domain tab XX which provides the labels, data types, and lengths by
variable. This data set can be used to test the \code{\link[=get_data_spec]{get_data_spec()}}
and \code{\link[=assign_meta_data]{assign_meta_data()}} functions.
}
\keyword{datasets}
