% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{trim_and_make_blanks_NA}
\alias{trim_and_make_blanks_NA}
\title{Standardize white space trimming and NA for character columns}
\usage{
trim_and_make_blanks_NA(tbl)
}
\arguments{
\item{tbl}{a data frame, the SDTM QC table}
}
\value{
a modified copy of the \code{tbl} data frame
}
\description{
Trims the white space on both sides and replaces blank values with \code{NA} for
all character columns.
}
\details{
This function should be applied as one of the first steps in the data process
to ensure consistent handling of strings.
}
