% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vd}
\alias{vd}
\title{Example EDC data for form/table 'VD'}
\format{
\subsection{'vd'}{

A data frame with 6 rows and 3 columns:
\describe{
\item{USUBJID}{Subject identifier}
\item{VISIT}{Visit name}
\item{VISITDTC}{Visit date}
}
}
}
\usage{
vd
}
\description{
This is an example data set to simulate raw EDC data from the 'VD' form/table
which contains visit date information by subject.
}
\keyword{datasets}
