% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xx_no_meta_data}
\alias{xx_no_meta_data}
\title{Example SDTM domain XX without meta data}
\format{
\subsection{'xx_no_meta_data'}{

A data frame with 10 rows and 11 columns:
\describe{
\item{STUDYID}{Study identifier}
\item{USUBJID}{Subject identifier}
\item{XXSEQ}{Sequence number}
\item{XXTESTCD}{Coded test name}
\item{XXTEST}{Test name}
\item{XXORRES}{Measurement in original units}
\item{XXBLFL}{Baseline flag}
\item{VISIT}{Visit name}
\item{EPOCH}{EPOCH}
\item{XXDTC}{Measurement date}
\item{XXDY}{Measurement day of study}
}
}
}
\usage{
xx_no_meta_data
}
\description{
This data set is used to test the \code{\link[=assign_meta_data]{assign_meta_data()}} function and contains
a fake SDTM domain XX but without label or lengths assigned to the column
attributes.
}
\keyword{datasets}
