% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{convert_to_script}
\alias{convert_to_script}
\title{Convert SDTM QC code from a .Rmd file to .R script}
\usage{
convert_to_script(dir, domain)
}
\arguments{
\item{dir}{string, the directory where the .Rmd file is and the .R file will
be written}

\item{domain}{string, the SDTM domain abbreviation}
}
\value{
nothing
}
\description{
Wraps \code{knitr::purl()} to create an .R script from a .Rmd file and also
auto-archives the .Rmd file to an \verb{[dir]/archive} sub-directory.
}
\details{
The function assumes the name of the .Rmd file is \code{"v_[domain].Rmd"} where
the domain is all lowercase. The resulting script will take the same name,
with a different extension (.R).
}
