% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_chars_and_dates}
\alias{format_chars_and_dates}
\title{Standardize formatting of date and character columns for QC SDTM tables}
\usage{
format_chars_and_dates(tbl)
}
\arguments{
\item{tbl}{a data frame, the SDTM QC table}
}
\value{
a modified copy of the \code{tbl} data frame
}
\description{
Converts all date columns to character class and replaces all \code{NA} values in
character/date columns with \code{""}.
}
\details{
This function should be applied as one of the last steps in the data process
but before \code{assign_meta_data()}.
}
