% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{summary_diffs}
\alias{summary_diffs}
\title{View \code{compareDF::compare_df()} summary}
\usage{
summary_diffs(comp)
}
\arguments{
\item{comp}{a list output from \code{compareDF::compare_df()}}
}
\value{
nothing
}
\description{
Prints two items, the first being the \code{change_summary} element of a
\code{compareDF::compare_df()} output list that shows a comparison of the total
number or records between the two data frames being compared. The second
printed item is a two column matrix that shows the number of \code{"+"} entries
(adds) and the number of \code{"-"} entries (dels) from a \code{compare_df()} output
list. The rows of the matrix are the columns in the two data frames being
compared. This function is used to summarize the output of
\code{compareDF::compare_df()} so the user can quickly see which columns have the
most differences.
}
