% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_tbl_to_xpt}
\alias{write_tbl_to_xpt}
\title{Write a SAS transport file (.xpt)}
\usage{
write_tbl_to_xpt(tbl, filename, dir)
}
\arguments{
\item{tbl}{a data frame to write}

\item{filename}{a string, the SDTM domain or supplemental domain name which
will become the file name and the name attribute of the transport file}

\item{dir}{a string, the directory to write to}
}
\value{
nothing
}
\description{
Writes a data frame to a SAS transport file (.xpt) named like \code{"[domain].xpt"}
}
\details{
Files will be written using version 5 .xpt files
}
